pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "frame"
version = "0.1"
description = "Floorplanning with RectilineAr ModulEs"
readme = "README.md"
license = { file = "LICENSE.txt" }
dependencies = [
    "ruamel.yaml",
    "gekko",
    "numpy",
    "matplotlib",
    "distinctipy",
    "Pillow",
    "python-sat",
    "tqdm",
    "networkx",
]

[project.optional-dependencies]
mypy = [
    "mypy",
    "types-Pillow",
    "types-setuptools"
]

jupyter = [
    "jupyterlab",
    "ipycanvas"
]

[project.urls]
homepage = "https://github.com/jordicf/FRAME"

[project.scripts]
frame = "tools.frame:main"

[tool.setuptools]
packages = [
    "frame",
    "frame.allocation",
    "frame.die",
    "frame.geometry",
    "frame.netlist",
    "frame.utils",
    "tools",
    "tools.draw",
    "tools.hello",
    "tools.netgen",
    "tools.spectral",
    "tools.force",
    "tools.glbfloor",
    "tools.rect",
    "tools.legalfloor",
    "tools.all"
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
plugins = "numpy.typing.mypy_plugin"
